cmd/gohandlers/int/pretty/sort: adds `Imports` to unify import spec sorting across subcommands
